Fix stuck-in-time bug
If the quantum is not a multiple of clock_cycle, after calling
simulate(num_cycles) there will be time left in the quantum. This
happens because num_cycles is rounded down. During the next iteration of
the while loop, num_cycles will be set to 0. This leads to a
not-increasing time, because simulate(0) is called which does not increase
local_time. To fix this behaviour, num_cycles is rounded up to ensure the full
quantum is used.

Signed-off-by: Nils Bosbach <bosbach@ice.rwth-aachen.de>