Fix usage of invalid pointer
The `__cxa__demangle` function gets a pointer to an allocated buffer
(`dmbuf`) and the size of allocated memory (`dmbufsz`) as paramters. If
the allocated size is not big enough to store the demangled name, more
memory is allocated using `realloc`. The allocated size is updated
(`dmbufsz`) and the pointer to the allocated memory is returned (`res`).
The pointer to the originally allocated memory (`dmbufsz`) might be
invalid if new memory is allocated using `realloc`.

Signed-off-by: Nils Bosbach <bosbach@ice.rwth-aachen.de>