Fix gic update bug
If the interrupt acknowledge procedure has not been finished when a new
interrupt is raised at the gic, this interrupt is not signaled to the CPU
immediately because of the pending one. Once the pending interrupt is
completely acknowledged (by writing to the EOI reg), the other pending
interrupt should be signaled to the CPU. This is done by calling update
after the acknowledged interrupt has been deactivated.

Signed-off-by: Nils Bosbach <bosbach@ice.rwth-aachen.de>