Sort regs of gdbserver target by regno
For the reg_read_all command, the gdb client expects the order of the
reg values to be the same as the order of the reg IDs. The
handle_reg_read_all function sends the values in the order of the
cpuregs vector of the target, which is ordered by the order of the
gdbarch feature struct.
To solve this issue, the regs vector is sorted based on the reg IDs.

Signed-off-by: Nils Bosbach <bosbach@ice.rwth-aachen.de>